refactor: split province overlay chrome and diplomacy row helpers (#3878)
Extract province overlay responsive chrome, trade screen tabs body, and
diplomacy row-builder helpers into library parts; document paths in
repo-and-packages.

Refs #3878
