+++ app/lib/features/game/screens/trade/trade_screen_tabs_body.dart
@@ -1,0 +1,60 @@
+// Two-tab body chrome for the World Market trade screen.
+
+part of 'trade_screen.dart';
+
+/// Two-tab body for the trade screen: Market (default) + Deal Book.
+///
+/// Hosts a [CtTabStrip] inside a [CtPanel] so the dark editorial-monocle
+/// surface mirrors the chrome already established for sibling
+/// full-screen feature surfaces (production, diplomacy). The Market tab
+/// now renders the interactive bid/offer/none + quantity stepper row
+/// sourced from [Game.worldMarketState] + `currentOrdersProvider`
+/// (Refs #2993 E5a + E5b); the Deal Book tab keeps the placeholder copy
+/// until the per-player ledger work for Refs #2993 E6 lands. The
+/// two-tab structure stays as the durable wireframe so the follow-up
+/// cargo indicator + priority dropdown + Deal Book ledger slices can
+/// swap each tab body in place without remounting the strip.
+class _TradeScreenTabsBody extends StatelessWidget {
+  const _TradeScreenTabsBody({
+    super.key,
+    required this.game,
+    required this.playerId,
+    required this.canEdit,
+    this.initialTabIndex = 0,
+  });
+
+  final Game game;
+  final String playerId;
+  final bool canEdit;
+  final int initialTabIndex;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(CtSpacing.l),
+      child: CtPanel(
+        padding: const EdgeInsets.all(CtSpacing.l),
+        child: CtTabStrip(
+          initialTabIndex: initialTabIndex,
+          tabLabels: const <String>[
+            TradeScreen.marketTabLabel,
+            TradeScreen.dealBookTabLabel,
+          ],
+          tabViews: <Widget>[
+            _MarketTabContent(
+              key: TradeScreen.marketTabBodyKey,
+              game: game,
+              playerId: playerId,
+              canEdit: canEdit,
+            ),
+            _DealBookTabContent(
+              key: TradeScreen.dealBookTabBodyKey,
+              game: game,
+              playerId: playerId,
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
